Conflicts:
	build_ontologies.py

+++ build_ontologies.py
@@ -11,11 +11,26 @@ from subprocess import Popen
 import shlex
 import urllib
 import shutil
-import time
 import glob
-import requests
 import fusekiutils
 import qudt4dt
+
+### Import stuff that's in our repo (in various places)
+### SO UGLY
+import inspect
+pathOfThisFile = inspect.getfile(inspect.currentframe())
+dirOfThisFile = os.path.split(pathOfThisFile)[0]
+
+sys.path.append(os.path.join(dirOfThisFile, 'modelica'))
+import MSL2OWL
+
+sys.path.append(os.path.join(dirOfThisFile, 'openMDAO'))
+import export_openMDAO_units
+
+sys.path.append(os.path.join(dirOfThisFile, 'requests'))
+import requests
+### End imports (and ugliness)
+
 
 p_ScriptPathRoot = os.path.dirname(__file__)
 ### Generate OpenMDAO Unit Ontology
@@ -44,13 +59,15 @@ print "complete"
 print ""
 
 ### start jena-fuseki and seed database ###
-print "Cleaning fuseki database"
 fuseki_data_path = os.getcwd() + '/fuseki-data'
-shutil.rmtree(fuseki_data_path)
-os.makedirs(fuseki_data_path)
-print "done"
+if os.path.exists(fuseki_data_path):
+    print "Cleaning fuseki database"
+    shutil.rmtree(fuseki_data_path)
+    print "done"
+    print ""
 
-print ""
+### Create fresh database folder
+os.makedirs(fuseki_data_path)
 
 fuseki_url = "http://localhost:3030"
 fuseki_upload_url = fuseki_url + '/qudt4dt/upload'
@@ -69,15 +86,14 @@ def LoadDirectoryOfOWLFiles(path):
         if filename.find('catalog-') != -1:
             continue
 
-        path_File = filename
         print 'Loading', filename
-        file_AsDict = {'file': open(path_File,'rb')}
-        r = requests.post(fuseki_upload_url, files=file_AsDict)
+        files = {'file': open(filename, 'rb')}
+        r = requests.post(fuseki_upload_url, files=files)
 
         if r.status_code != 200:
             print "EXCEPTION loading DB: ", r.status_code
 
-def CreateOWLFilesFromCSV(sourceFilePath ,objFilePath):
+def CreateOWLFilesFromCSV(sourceFilePath, objFilePath):
     s_file = re.search('[^/]+?\.csv',sourceFilePath).group()
     o_file = re.search('[^/]+?\.xml',objFilePath).group()
     print 'coverting %s to %s' %(s_file, o_file)
@@ -92,6 +108,7 @@ try:
     LoadDirectoryOfOWLFiles('qudt-owl')
     LoadDirectoryOfOWLFiles('modelica')
     LoadDirectoryOfOWLFiles('ontologies')
+    LoadDirectoryOfOWLFiles('openMDAO')
     print "done"
 
 finally:
